refactor: command 패키지 생성
