+++ src/main/java/chess/domain/command/Command.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.command;
 
 import java.util.Arrays;
 import java.util.Objects;
