+++ src/main/java/chess/domain/command/FirstCommand.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.command;
 
 import java.util.Objects;
